module_requirements.txt
mako
virustotal-api
ares
netifaces
pexpect
requests
